backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
alembic==1.14.1
psycopg[binary]==3.2.13
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
redis==5.2.1
rq==2.3.3
boto3==1.35.99
sentry-sdk[fastapi]==2.19.2
email-validator==2.2.0
python-multipart==0.0.20
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
pypdf==5.4.0
